requirements_travis.txt
pandas>=0.23.4
numpy>=1.15.2
scipy>=1.1.0
matplotlib==2.2.3
biopython>=1.72
pathos>=0.2.2.1
pytest>=3.8.2
twine>=1.12.1
olga>=1.0.2
